requirements.txt
python-dotenv
firebase-admin
discord.py